Add cachet plugin
This pluin provides a way to receive service outage notifications as detected in
Cachet service. It uses cachet API to get all failed services.

For right now the plugin does not support API paging so if there are a lot of
outages some notifications will be missed. It also relies on dedup plugin to
deduplicate alert messages which would otherwise be reported every minute.